refactor: :truck: firebaseをディレクトリではなく単一のファイルに変更
