+++ packages/app/src/shared/libs/firebase.ts
